Initial support for custom kernel
This change allows for selection of a specific kernel for the bootstrap
iso and adds initial Photon 2.0 support.

Outstanding:
1. allow each config to supply it's own kernel parameters in isolinux.cfg
2. changes to select the "ci" repo when building in CI
3. Photon2.0 based appliance doesn't start
4. allow each config to supply a custom set of packages to use

The last point is specifically in aid of RHEL6.9 support which doesn't
seem to function well with systemd. It's possible that there's tweaks to
be made that will address this, but testing with running systemd directly
in a booted bootstrap ISO has not made much progress.
It's unable to find cgroup config which isn't present in /sys/fs/cgroup.

As such I've been considering switching back to udevd for initial system
configuration.

I've been using the following to build the test ISO:
export DEBUG=1
export CUSTOM=rhel69
export CUSTOM_RPM=`pwd`/kernel-2.6.32-696.10.2.el6.x86_64.rpm
sudo -E make bootstrap-custom

Centos6.9 bootstrap initial work

Updates the iso-base-custom target to use CUSTOM name to select repo.
Makes changes to construct initial base.

Not yet tried running bootstrap-staging.

Mostly functioning custom iso build

Gets rid of the photon2.0 specific targets and makes it just another repo.
photon-2.0 base bootstrap is untested currently.

Adds a basic breakout of the live init process after bootstrap boot and
makes the use of entropy daemon and iptables a little more generalized.

I do not think this is a good final form and IPTABLES CONFIG IS NOT
CORRECTLY PERFORMED AS YET.

Currently building custom ISO via:
CUSTOM=centos-6.9 sudo -E make bootstrap-custom

WIP: moves to toybox but broken

yum does not support rich dependencies so we will need to move to tdnf as
a package manager during the build in order to adopt photon 2.0 reliably.

WIP: photon-1.0 based passthrough config

This adds a repo configured for the vanilla kernel and installs
linux-tools so that we have the necessary udb drivers available for a
pass-through deployment

Plenty of work still outstanding and various hacks/dependencies that need
revisiting.

Photon-2 seems to have broken since the last commit on this branch due to
an sqllite dependency of some kind. Likely requires a move to tdnf instead
of yum in the install scripts to leverage the granular dependencies.

Built with:
export CUSTOM=photon-1.0-passthrough
sudo -E make bootstrap-custom appliance vic-machine-linux

Deployed with:
vic-create --debug=2 --bi=bootstrap-photon-1.0-passthrough.iso

Notes:
iptables - still not working right, but configured for open by default
until it's addressed

Tested with alpine and ubuntu

Add exe permissions to scripts

+++ isos/bootstrap.sh
@@ -48,7 +48,7 @@ echo "Usage: $0 -p staged-package(tgz) -b binary-dir -d <activates debug when se
 exit 1
 }
 
-while getopts "p:b:d:" flag
+while getopts "p:b:d:o:" flag
 do
     case $flag in
 
@@ -64,6 +64,10 @@ do
         d)
             # Optional. directs script to make a debug iso instead of a production iso.
             debug="$OPTARG"
+            ;;
+        o)
+            # Optional. Name of the generated ISO
+            ISONAME="$OPTARG"
             ;;
         *)
 
@@ -88,13 +92,16 @@ PKGDIR=$(mktemp -d)
 
 unpack $package $PKGDIR
 
+# load the repo to use from the package if not explicit in env
+REPODIR="$DIR/base/repos/${REPO:-$(cat $PKGDIR/repo.cfg)}/"
+
 #selecting the init script as our entry point.
 if [ -v debug ]; then
-    export ISONAME="bootstrap-debug.iso"
+    export ISONAME=${ISONAME:-bootstrap-debug.iso}
     cp ${DIR}/bootstrap/bootstrap.debug $(rootfs_dir $PKGDIR)/bin/bootstrap
     cp ${BIN}/rpctool $(rootfs_dir $PKGDIR)/sbin/
 else
-    export ISONAME="bootstrap.iso"
+    export ISONAME=${ISONAME:-bootstrap.iso}
     cp ${DIR}/bootstrap/bootstrap $(rootfs_dir $PKGDIR)/bin/bootstrap
 fi
 
@@ -102,20 +109,33 @@ fi
 cp ${BIN}/tether-linux $(rootfs_dir $PKGDIR)/bin/tether
 cp ${BIN}/unpack $(rootfs_dir $PKGDIR)/bin/unpack
 
-# kick off our components at boot time
-mkdir -p $(rootfs_dir $PKGDIR)/etc/systemd/system/vic.target.wants
-cp ${DIR}/bootstrap/tether.service $(rootfs_dir $PKGDIR)/etc/systemd/system/
-cp ${DIR}/appliance/vic.target $(rootfs_dir $PKGDIR)/etc/systemd/system/
-ln -s /etc/systemd/system/tether.service $(rootfs_dir $PKGDIR)/etc/systemd/system/vic.target.wants/
-ln -sf /etc/systemd/system/vic.target $(rootfs_dir $PKGDIR)/etc/systemd/system/default.target
+if [ -d $(rootfs_dir $PKGDIR)/etc/systemd ]; then
+    echo "Preparing systemd for bootstrap"
 
-# disable networkd given we manage the link state directly
-rm -f $(rootfs_dir $PKGDIR)/etc/systemd/system/multi-user.target.wants/systemd-networkd.service
-rm -f $(rootfs_dir $PKGDIR)/etc/systemd/system/sockets.target.wants/systemd-networkd.socket
+    # copy in sysv-init entry script
+    cp ${DIR}/bootstrap/systemd-init $(rootfs_dir $PKGDIR)/bin/init
 
-# do not use the systemd dhcp client
-rm -f $(rootfs_dir $PKGDIR)/etc/systemd/network/*
-cp ${DIR}/base/no-dhcp.network $(rootfs_dir $PKGDIR)/etc/systemd/network/
+    # kick off our components at boot time
+    mkdir -p $(rootfs_dir $PKGDIR)/etc/systemd/system/vic.target.wants
+    cp ${DIR}/bootstrap/tether.service $(rootfs_dir $PKGDIR)/etc/systemd/system/
+    cp ${DIR}/appliance/vic.target $(rootfs_dir $PKGDIR)/etc/systemd/system/
+    ln -s /etc/systemd/system/tether.service $(rootfs_dir $PKGDIR)/etc/systemd/system/vic.target.wants/
+    ln -sf /etc/systemd/system/vic.target $(rootfs_dir $PKGDIR)/etc/systemd/system/default.target
+
+    # disable networkd given we manage the link state directly
+    rm -f $(rootfs_dir $PKGDIR)/etc/systemd/system/multi-user.target.wants/systemd-networkd.service
+    rm -f $(rootfs_dir $PKGDIR)/etc/systemd/system/sockets.target.wants/systemd-networkd.socket
+
+    # do not use the systemd dhcp client
+    rm -f $(rootfs_dir $PKGDIR)/etc/systemd/network/*
+    cp ${DIR}/base/no-dhcp.network $(rootfs_dir $PKGDIR)/etc/systemd/network/
+else
+    echo "Preparing systemV init for bootstrap"
+    
+    # copy in sysv-init entry script
+    cp ${DIR}/bootstrap/sysv-init $(rootfs_dir $PKGDIR)/bin/init
+fi
+cp ${REPODIR}/init.sh $(rootfs_dir $PKGDIR)/bin/repoinit
 
 # compute the size of the target tempfs,
 # the list of directories/files in ${tempfs_target_list} should
@@ -131,4 +151,5 @@ echo Total tempfs size: ${size}
 echo "${tempfs_target_list[@]}" > $(rootfs_dir $PKGDIR)/.tempfs_list
 echo ${size} > $(rootfs_dir $PKGDIR)/.tempfs_size
 
-generate_iso $PKGDIR $BIN/$ISONAME /lib/systemd/systemd
+INIT=$(cat $REPODIR/init.cfg | awk '/^[^#]/{print}')
+generate_iso $PKGDIR $BIN/$ISONAME $INIT
